refactor(beast): 重构灵兽实体和福地服务中的灵兽管理逻辑
- 优化 canFight 方法，添加恢复时间检查和状态清理
- 添加 levelCap 判空保护，避免 NPE 异常
- 重构 recalculateAttributes 方法，统一属性计算公式
- 新增 getQualityMultiplier 方法，标准化品质倍率计算
- 移除 CombatService 中的战斗觉醒实现，统一调用 FudiService
- 优化 FudiService 中的灵兽属性计算方法
- 重构灵兽孵化、进化、突破等核心业务逻辑
- 统一灵兽状态同步机制，移除重复的状态设置代码
- 优化灵兽出战和召回逻辑，提升性能表现
- 清理多余导入和注释，提升代码可读性

+++ src/main/java/top/stillmisty/xiantao/domain/land/vo/PenCellVO.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class PenCellVO {
     private Integer cellId;
     private Integer cellLevel;
-    private Integer beastId;
+    private Long beastId;
     private String beastName;
     private Integer tier;
     private String quality;
